+++ src/main.go
@@ -12,7 +12,7 @@ import (
 	"github.com/tminaorg/brzaguza/src/cache/nocache"
 	"github.com/tminaorg/brzaguza/src/cache/pebble"
 	"github.com/tminaorg/brzaguza/src/cache/redis"
-	"github.com/tminaorg/brzaguza/src/climode"
+	"github.com/tminaorg/brzaguza/src/cli"
 	"github.com/tminaorg/brzaguza/src/config"
 	"github.com/tminaorg/brzaguza/src/logger"
 	"github.com/tminaorg/brzaguza/src/router"
@@ -22,26 +22,26 @@ func main() {
 	mainTimer := time.Now()
 
 	// parse cli arguments
-	setupCli()
+	cliFlags := cli.Setup()
 
-	_, stopProfiler := runProfiler()
+	_, stopProfiler := runProfiler(&cliFlags)
 	defer stopProfiler()
 
 	// signal interrupt (CTRL+C)
 	ctx, _ := signal.NotifyContext(context.Background(), os.Interrupt, syscall.SIGINT, syscall.SIGTERM)
 
 	// configure logging
-	logger.Setup(cli.Log, cli.Verbosity)
+	logger.Setup(cliFlags.Log, cliFlags.Verbosity)
 
 	// load config file
 	conf := config.New()
-	conf.Load(cli.Config, cli.Log)
+	conf.Load(cliFlags.Config, cliFlags.Log)
 
 	// cache database
 	var db cache.DB
 	switch conf.Server.Cache.Type {
 	case "pebble":
-		db = pebble.New(cli.Config)
+		db = pebble.New(cliFlags.Config)
 	case "redis":
 		db = redis.New(ctx, conf.Server.Cache.Redis)
 	default:
@@ -50,13 +50,13 @@ func main() {
 	}
 
 	// startup
-	if cli.Cli {
-		climode.Run(cli.Query, cli.MaxPages, cli.Visit, cli.Silent, db, conf)
+	if cliFlags.Cli {
+		cli.Run(cliFlags, db, conf)
 	} else {
-		if rw, err := router.New(conf, cli.Verbosity); err != nil {
+		if rw, err := router.New(conf, cliFlags.Verbosity); err != nil {
 			log.Error().Msgf("Failed creating a router: %v", err)
 		} else {
-			rw.Start(ctx, db, cli.ServeProfiler)
+			rw.Start(ctx, db, cliFlags.ServeProfiler)
 		}
 	}
